fix powblock

+++ MarioBaseProject/constants.h
@@ -1,13 +1,13 @@
 #pragma once
-const int MOVEMENTSPEED = 100.0f;
-const int GRAVITY = 260.0f;
-const int JUMP_FORCE_DECREMENT = 400.0f;
-const int INITIAL_JUMP_FORCE = 300.0f;
+const float MOVEMENTSPEED = 100.0f;
+const float GRAVITY = 260.0f;
+const float JUMP_FORCE_DECREMENT = 400.0f;
+const float INITIAL_JUMP_FORCE = 300.0f;
 const int MAP_HEIGHT = 13;
 const int MAP_WIDTH = 16;
 const int TILE_WIDTH = 32;
 const int TILE_HEIGHT = 32;
-const int SHAKE_DURATION = 0.25f;
+const float SHAKE_DURATION = 0.25f;
 
 //screen Dimensions
 #define SCREEN_WIDTH 512
